refactor: 계층 분리, 명칭 변경
