+++ src/main/java/com/hyunec/cosmicbaseballinit/batter/domain/BattingResult.java
@@ -1,0 +1,5 @@
+package com.hyunec.cosmicbaseballinit.batter.domain;
+
+public enum BattingResult {
+    STRIKE, BALL, HIT
+}
