+++ src/main/java/com/hyunec/cosmicbaseballinit/batter/service/BattingStrategy.java
@@ -1,0 +1,9 @@
+package com.hyunec.cosmicbaseballinit.batter.service;
+
+import com.hyunec.cosmicbaseballinit.batter.domain.BattingResult;
+
+public interface BattingStrategy {
+
+    BattingResult generateBatting();
+
+}
