--- src/main/java/com/hyunec/cosmicbaseballinit/batting/BattingResults.java
@@ -1,5 +1,0 @@
-package com.hyunec.cosmicbaseballinit.batting;
-
-public enum BattingResults {
-    STRIKE, BALL, HIT
-}
